Rename reserved keyword "async" for Python 3.7 compatibility. Introduce pre-commit hooks to avoid such.
