+++ .pre-commit-config.yaml
@@ -7,14 +7,9 @@
     -   id: check-added-large-files
     -   id: debug-statements
     -   id: end-of-file-fixer
--   repo: git://github.com/FalconSocial/pre-commit-python-sorter
-    sha: 5991d2aea26858d3c3538e0b4e09255b6b99413e
+        exclude: '^.+?\.json$'
+-   repo: https://github.com/asottile/reorder_python_imports
+    sha: v0.3.5
     hooks:
-    -   id: python-import-sorter
-        args:
-        - --silent-overwrite
--   repo: git://github.com/pre-commit/mirrors-eslint
-    sha: v3.14.0
-    hooks:
-    -   id: eslint
-        additional_dependencies: [ 'eslint', 'eslint-plugin-html', 'eslint-config-airbnb', 'eslint-plugin-import', 'eslint-plugin-jsx-a11y']
+    -   id: reorder-python-imports
+        language_version: python2.7
