+++ setup.cfg
@@ -1,0 +1,20 @@
+[wheel]
+universal = 1
+
+[aliases]
+test = pytest
+
+[flake8]
+max-line-length = 160
+max-complexity = 10
+
+# Ignore non-PEP8-compliant rules so that the Black formatter can be used
+ignore = E203,W503
+
+[isort]
+atomic = true
+multi_line_output = 5
+line_length = 160
+indent = '    '
+combine_as_imports = true
+skip = wsgi.py,docs,env,cli.py,test,.eggs,build,setup.py
